+++ order/src/main/java/order/dto/CreateOrderRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.order;
+package order.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
