+++ order/src/main/java/order/exception/OrderTableNotFoundException.java
@@ -1,6 +1,6 @@
-package kitchenpos.exception.order;
+package order.exception;
 
-import kitchenpos.exception.common.NotFoundException;
+import common.exception.NotFoundException;
 
 public class OrderTableNotFoundException extends NotFoundException {
     private static final String RESOURCE = "주문 테이블";
