+++ settings.gradle
@@ -3,3 +3,4 @@ include 'menu-groups'
 include 'common'
 include 'product'
 include 'menu'
+include 'order'
