--- src/main/java/kitchenpos/dto/order/ChangeOrderStatusResponse.java
@@ -1,4 +1,0 @@
-package kitchenpos.dto.order;
-
-public class ChangeOrderStatusResponse {
-}
